move functions to new file
